rename to DualInlinePackage
